requirements.txt
pydantic>=2.0
pydantic-settings>=2.0
pydantic-ai>=0.1
fastapi>=0.110
uvicorn[standard]>=0.27
httpx>=0.27
streamlit>=1.30
sqlalchemy[asyncio]>=2.0
aiosqlite>=0.19
google-auth-oauthlib>=1.0
google-api-python-client>=2.0
beautifulsoup4>=4.12
python-dotenv>=1.0
chromadb>=0.4
passlib[bcrypt]>=1.7
python-jose[cryptography]>=3.3
